Rand01 is not float

+++ Utils.h
@@ -10,8 +10,8 @@ inline T Clamp01(T v) {
     return v < T(0) ? T(0) : v > T(1) ? T(1) : v;
 }
 
-inline double Rand01() {
-    return (double)rand() / (double)RAND_MAX;
+inline float Rand01() {
+    return (float)rand() / (float)RAND_MAX;
 }
 
 inline int ToByte(float v) {
